EMP-21046 - QRScannerViewModel
